+++ CHANGELOG.md
@@ -13,10 +13,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `stac_version` to the default set of fields returned when the `fields` parameter
   is an empty value.
 
+## Removed
+
+- POST /aggregate endpoints were removed, as they didn't work correctly.
+
 ## Added
 
 - Added support for `/aggregations`, `/collections/{collectionId}/aggregations`, and
   `/collections/{collectionId}/aggregate` endpoints.
+- Added grid_geohex_frequency, grid_geohash_frequency, and grid_geotile_frequency aggregations
 
 ## Fixed